# apps/api/docs/swagger.yaml
definitions:
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  dto.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  dto.TokenRefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  dto.UserActivityItem:
    properties:
      description:
        type: string
      id:
        type: string
      occurred_at:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  dto.UserProgress:
    properties:
      completed_lessons:
        type: integer
      completion_rate:
        type: number
      course_id:
        type: string
      course_title:
        type: string
      last_activity_at:
        type: string
      total_lessons:
        type: integer
    type: object
  dto.UserStats:
    properties:
      average_score:
        type: number
      completed_courses:
        type: integer
      current_streak:
        type: integer
      total_study_minutes:
        type: integer
    type: object
  dto.VerifyEmailRequest:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
...
  model.Category:
    properties:
      children:
        items:
          $ref: '#/definitions/model.Category'
        type: array
      children_count:
        description: Computed field
        type: integer
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      description_translations:
        $ref: '#/definitions/model.TranslatableContent'
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      name_translations:
        $ref: '#/definitions/model.TranslatableContent'
      order:
        type: integer
      parent:
        allOf:
        - $ref: '#/definitions/model.Category'
        description: Relationships
      parent_id:
        type: integer
      slug:
        type: string
      status:
        $ref: '#/definitions/model.CategoryStatus'
      type:
        $ref: '#/definitions/model.CategoryType'
      updated_at:
        type: string
    type: object
  model.CategoryStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - CategoryStatusActive
    - CategoryStatusInactive
  model.CategoryType:
    enum:
    - blog
    - header
    - footer
    - sidebar
    type: string
    x-enum-varnames:
    - CategoryTypeBlog
    - CategoryTypeHeader
    - CategoryTypeFooter
    - CategoryTypeSidebar
  model.Department:
    properties:
      children:
        items:
          $ref: '#/definitions/model.Department'
        type: array
      code:
        description: e.g., "sales", "editorial", "it"
        type: string
      created_at:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_system:
        type: boolean
      manager_id:
        description: Department manager (User ID would be UUID in real case)
        type: integer
      module:
        allOf:
        - $ref: '#/definitions/model.Module'
        description: Relationships
      module_id:
        type: integer
      name:
        type: string
      parent:
        $ref: '#/definitions/model.Department'
      parent_id:
        description: For hierarchical departments
        type: integer
      services:
        items:
          $ref: '#/definitions/model.Service'
        type: array
      updated_at:
        type: string
    type: object
  model.Media:
    properties:
      alt:
        description: Alt text for images
        type: string
      bucket:
        description: MinIO bucket
        type: string
      caption:
        type: string
      created_at:
        type: string
      description:
        type: string
      duration:
        description: For videos/audio (in seconds)
        type: integer
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      height:
        description: For images/videos
        type: integer
      id:
        type: integer
      mime_type:
        type: string
      object_key:
        description: MinIO object key
        type: string
      original_name:
        type: string
      posts:
        items:
          $ref: '#/definitions/model.Post'
        type: array
      tags:
        description: JSON array of tags
        type: string
      thumbnail:
        description: Thumbnail path for videos
        type: string
      type:
        $ref: '#/definitions/model.MediaType'
      updated_at:
        type: string
      uploaded_by:
        type: string
      uploader:
        allOf:
        - $ref: '#/definitions/model.User'
        description: Relationships
      url:
        description: Public URL
        type: string
      width:
        description: For images/videos
        type: integer
    type: object
  model.MediaType:
    enum:
    - image
    - video
    - document
    - audio
    - other
    type: string
    x-enum-varnames:
    - MediaTypeImage
    - MediaTypeVideo
    - MediaTypeDocument
    - MediaTypeAudio
    - MediaTypeOther
  model.Module:
    properties:
      code:
        description: e.g., "crm", "content", "admin"
        type: string
      color:
        type: string
      created_at:
        type: string
      departments:
        description: Relationships
        items:
          $ref: '#/definitions/model.Department'
        type: array
      description:
        type: string
      display_name:
        type: string
      icon:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_system:
        description: System modules cannot be deleted
        type: boolean
      name:
        type: string
      order:
        type: integer
      updated_at:
        type: string
    type: object
  model.Permission:
    properties:
      action:
        description: e.g., "create", "read", "update", "delete"
        type: string
      created_at:
        type: string
      department:
        description: Department level
        type: string
      description:
        type: string
      id:
        type: integer
      module:
        description: e.g., "crm", "content", "admin"
        type: string
      resource:
        description: e.g., "users", "posts", "customers"
        type: string
      roles:
        description: Relationships
        items:
          $ref: '#/definitions/model.Role'
        type: array
      service:
        description: Service level within department
        type: string
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
  model.Post:
    properties:
      author:
        allOf:
        - $ref: '#/definitions/model.User'
        description: Relationships
      author_id:
        type: string
      author_name:
        type: string
      categories:
        description: JSON array of category IDs
        type: string
      category_ids:
        items:
          type: integer
        type: array
      category_objects:
        items:
          $ref: '#/definitions/model.Category'
        type: array
      comment_count:
        type: integer
      content:
        type: string
      content_translations:
        $ref: '#/definitions/model.TranslatableContent'
      created_at:
        type: string
      excerpt:
        type: string
      excerpt_translations:
        $ref: '#/definitions/model.TranslatableContent'
      featured_image:
        type: string
      id:
        type: integer
      language:
        type: string
      like_count:
        type: integer
      media:
        items:
          $ref: '#/definitions/model.Media'
        type: array
      meta_description:
        type: string
      meta_keywords:
        type: string
      meta_title:
        type: string
      published_at:
        type: string
      scheduled_at:
        type: string
      slug:
        type: string
      slug_translations:
        $ref: '#/definitions/model.TranslatableContent'
      status:
        $ref: '#/definitions/model.PostStatus'
      table_of_contents:
        description: JSON array of TOC items
        type: string
      tags:
        description: JSON array of tags
        type: string
      title:
        type: string
      title_translations:
        $ref: '#/definitions/model.TranslatableContent'
      updated_at:
        type: string
      view_count:
        type: integer
    type: object
  model.PostStatus:
    enum:
    - draft
    - scheduled
    - published
    - archived
    type: string
    x-enum-varnames:
    - PostStatusDraft
    - PostStatusScheduled
    - PostStatusPublished
    - PostStatusArchived
  model.Role:
    properties:
      children:
        items:
          $ref: '#/definitions/model.Role'
        type: array
      created_at:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: integer
      is_system:
        description: System roles cannot be deleted
        type: boolean
      level:
        $ref: '#/definitions/model.RoleLevel'
      name:
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/model.Role'
        description: Relationships
      parent_id:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/model.Permission'
        type: array
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
  model.RoleLevel:
    enum:
    - organization
    - department
    - service
    - action
    type: string
    x-enum-varnames:
    - RoleLevelOrganization
    - RoleLevelDepartment
    - RoleLevelService
    - RoleLevelAction
  model.Service:
    properties:
      code:
        description: e.g., "user_management", "post_management"
        type: string
      created_at:
        type: string
      department:
        allOf:
        - $ref: '#/definitions/model.Department'
        description: Relationships
      department_id:
        type: integer
      description:
        type: string
      display_name:
        type: string
      endpoint:
        description: API endpoint prefix
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_system:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.TranslatableContent:
    additionalProperties:
      type: string
    type: object
  model.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      department:
        allOf:
        - $ref: '#/definitions/model.Department'
        description: Relationships
      department_id:
        description: Foreign key to departments table
        type: integer
      email:
        type: string
      email_verified:
        type: boolean
      email_verified_at:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_login_at:
        type: string
      last_login_ip:
        type: string
      last_name:
        type: string
      media:
        items:
          $ref: '#/definitions/model.Media'
        type: array
      permissions:
        items:
          $ref: '#/definitions/model.Permission'
        type: array
      phone:
        type: string
      position:
        description: Job title/position
        type: string
      posts:
        items:
          $ref: '#/definitions/model.Post'
        type: array
      roles:
        items:
          $ref: '#/definitions/model.Role'
        type: array
      status:
        $ref: '#/definitions/model.UserStatus'
      tenantID:
        type: string
      two_factor_enabled:
        type: boolean
      updated_at:
        type: string
    type: object
  model.UserStatus:
    enum:
    - active
    - inactive
    - suspended
    - pending
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusInactive
    - UserStatusSuspended
    - UserStatusPending
...
  response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      request_id:
        type: string
      success:
        type: boolean
    type: object
  response.ErrorBody:
    properties:
      code:
        type: integer
      fields:
        additionalProperties:
          type: string
        description: validation field errors
        type: object
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  service.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
...
  service.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  service.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
...
info:
  contact: {}
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Forgot password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Request password reset
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/service.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Login with email and password
      tags:
      - auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Logout current user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TokenRefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/service.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Register a new user
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Reset password
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: Verify email payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Verify email address
      tags:
      - auth
...
  /users/activities:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by title or description
        in: query
        name: q
        type: string
      - description: Filter by activity type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UserActivityItem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get current user activities
      tags:
      - users
...
  /users/progress:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UserProgress'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get current user course progress
      tags:
      - users
  /users/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserStats'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get current user stats
      tags:
      - users
...
swagger: "2.0"